feat : user-event testing(cont.)

+++ controllers/user_event.js
@@ -9,6 +9,7 @@ const {
   AnswerQuiz,
   sequelize,
 } = require("../models");
+const event = require("../routers/event");
 
 module.exports = class userEventController {
   static getAllEvents = async (req, res, next) => {
@@ -165,6 +166,8 @@ module.exports = class userEventController {
   static deleteEvent = async (req, res, next) => {
     try {
       const { id } = req.params;
+      const user_event = User_Event.findByPk(id);
+      if (user_event) throw ({ name: "Data not found!" })
       await User_Event.destroy({
         where: { id },
       });
